Extract component state and behaviour to component hook
