+++ apps/myaccount/src/hooks/use-push-authenticator.ts
@@ -16,6 +16,134 @@
  * under the License.
  */
 
+import React, { useEffect, useState } from "react";
+import { useTranslation } from "react-i18next";
+import { useDispatch } from "react-redux";
+import { Dispatch } from "redux";
+import useGetPushAuthRegisteredDevices from "./use-get-push-auth-registered-devices";
+import { deletePushAuthRegisteredDevice, initPushAuthenticatorQRCode } from "../api/multi-factor-push";
+import { AlertLevels } from "../models/alert";
+import { HttpResponse } from "../models/api";
+import { addAlert } from "../store/actions/global";
+
+/**
+ * Custom hook to handle the component state and behaviour of {@Link PushAuthenticator} component
+ */
 export default function usePushAuthenticator () {
-    
+    const {
+        data: registeredDeviceList,
+        isLoading:isRegisteredDeviceListLoading,
+        error: registeredDeviceListFetchError,
+        mutate: updateRegisteredDeviceList
+    } = useGetPushAuthRegisteredDevices();
+
+    const dispatch: Dispatch = useDispatch();
+
+    const [ isLoading, setIsLoading ] = useState<boolean>(false);
+    const [ isConfigPushAuthenticatorModalOpen, setIsConfigPushAuthenticatorModalOpen ] = useState<boolean>(false);
+    const [ qrCode, setQrCode ] = useState<string>(null);
+
+    const { t } = useTranslation();
+
+    const translateKey: string = "myAccount:components.mfa.pushAuthenticatorApp.";
+
+    useEffect(() => {
+        if (registeredDeviceListFetchError && !isRegisteredDeviceListLoading) {
+            dispatch(addAlert({
+                description: t(translateKey +
+                        "notifications.updateAuthenticatorError.error.description", {
+                    error: registeredDeviceListFetchError?.message
+                }),
+                level: AlertLevels.ERROR,
+                message: t(translateKey + "notifications.updateAuthenticatorError.error.message")
+            }));
+        }
+    }, [ isRegisteredDeviceListLoading, registeredDeviceListFetchError ]);
+
+    /**
+     * Initiate the push authenticator configuration flow.
+     */
+    const initPushAuthenticatorRegFlow = () => {
+        setIsLoading(true);
+
+        initPushAuthenticatorQRCode()
+            .then((response: any) => {
+                const qrCode: string = window.btoa(JSON.stringify(response?.data));
+
+                setIsConfigPushAuthenticatorModalOpen(true);
+                setQrCode(qrCode);
+            })
+            .catch((error: any) => {
+                dispatch(addAlert({
+                    description: t(translateKey + "notifications.initError.error.description", {
+                        error
+                    }),
+                    level: AlertLevels.ERROR,
+                    message: t(translateKey + "notifications.initError.error.message")
+                }));
+            })
+            .finally(() => {
+                setIsLoading(false);
+            });
+    };
+
+    /**
+     * Handle cancelling push authenticator configuration flow.
+     */
+    const handlePushAuthenticatorInitCancel = () => {
+        setIsConfigPushAuthenticatorModalOpen(false);
+        setQrCode(null);
+    };
+
+    /**
+     * Handle Push Authenticator configuration flow.
+     *
+     * @param event - Form submit event.
+     * @param isRegenerated - Whether the push authenticator QR code is regenerated or not.
+     */
+    const handlePushAuthenticatorSetupSubmit = (event: React.FormEvent<HTMLFormElement>): void => {
+        event.preventDefault();
+        updateRegisteredDeviceList();
+        setIsConfigPushAuthenticatorModalOpen(false);
+    };
+
+    /**
+     * Delete a device registered for push authentication.
+     *
+     * @param deviceId - unique id of the registered device
+     */
+    const deleteRegisteredDevice = (deviceId: string): void => {
+        setIsLoading(true);
+        deletePushAuthRegisteredDevice(deviceId).then(
+            (_res: HttpResponse) => {
+                updateRegisteredDeviceList();
+                dispatch(addAlert({
+                    description: "successfully deleted",
+                    level: AlertLevels.SUCCESS,
+                    message: "delete success"
+                }));
+            }
+        ).catch((_err: any) => {
+            dispatch(addAlert({
+                description: "error occurred when deleting the registered device",
+                level: AlertLevels.ERROR,
+                message: "delete error"
+            }));
+        }).finally(() => {
+            setIsLoading(false);
+        });
+    };
+
+    return {
+        deleteRegisteredDevice,
+        handlePushAuthenticatorInitCancel,
+        handlePushAuthenticatorSetupSubmit,
+        initPushAuthenticatorRegFlow,
+        isConfigPushAuthenticatorModalOpen,
+        isLoading,
+        isRegisteredDeviceListLoading,
+        qrCode,
+        registeredDeviceList,
+        translateKey
+    };
 }
